+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Client: Loop using a more fancy restart loop that uses a variable wait time and max restarts
 - Server: Do not spam log by not logging EOF, TCP reader or clean exits
 - Server: Set Upgrade and Connection headers sooner
+- Client+Server: Compare headers case insensitive
 - General: Rename Proxyguard to ProxyGuard
 
 # 0.4.0